Update caption time fields to use float for better precision

+++ zeeguu/core/model/caption.py
@@ -10,8 +10,8 @@ class Caption(db.Model):
     video = db.relationship("Video", back_populates="captions")
     video_id = db.Column(db.Integer, db.ForeignKey("video.id"))
 
-    time_start = db.Column(db.Integer)
-    time_end = db.Column(db.Integer)
+    time_start = db.Column(db.Float)
+    time_end = db.Column(db.Float)
 
     text_id = db.Column(db.Integer, db.ForeignKey(NewText.id))
     text = db.relationship(NewText, foreign_keys="Caption.text_id")
